rename : change rootProject Name to motus
