+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'rehab'
+rootProject.name = 'motus'
