--- src/main/java/com/hallym/rehab/RehabApplication.java
@@ -1,15 +1,0 @@
-package com.hallym.rehab;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@SpringBootApplication
-@EnableJpaAuditing
-public class RehabApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(RehabApplication.class, args);
-	}
-
-}
