--- src/main/java/com/hallym/rehab/domain/video/entity/Category.java
@@ -1,5 +1,0 @@
-package com.hallym.rehab.domain.video.entity;
-
-public enum Category {
-    ARM, SHOULDER, KNEE, THIGH
-}
