--- src/main/java/com/hallym/rehab/domain/video/entity/Position.java
@@ -1,5 +1,0 @@
-package com.hallym.rehab.domain.video.entity;
-
-public enum Position {
-    STANDING, SITTING, LYING
-}
