--- src/main/java/com/hallym/rehab/domain/video/repository/VideoRepository.java
@@ -1,8 +1,0 @@
-package com.hallym.rehab.domain.video.repository;
-
-import com.hallym.rehab.domain.video.entity.Video;
-import com.hallym.rehab.domain.video.service.videoSearch.VideoSearch;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface VideoRepository extends JpaRepository<Video, Long>, VideoSearch {
-}
