--- src/main/java/com/hallym/rehab/domain/video/service/videoSearch/VideoSearch.java
@@ -1,9 +1,0 @@
-package com.hallym.rehab.domain.video.service.videoSearch;
-
-import com.hallym.rehab.domain.video.dto.VideoResponseDTO;
-import com.hallym.rehab.domain.video.dto.pagedto.VideoPageRequestDTO;
-import org.springframework.data.domain.Page;
-
-public interface VideoSearch {
-    Page<VideoResponseDTO> search(VideoPageRequestDTO videoPageRequestDTO);
-}
