--- src/main/java/com/hallym/rehab/global/config/S3Client.java
@@ -1,35 +1,0 @@
-package com.hallym.rehab.global.config;
-
-import com.amazonaws.auth.AWSStaticCredentialsProvider;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.client.builder.AwsClientBuilder;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class S3Client {
-
-    @Value("${cloud.aws.credentials.access-key}")
-    private String accessKey;
-
-    @Value("${cloud.aws.credentials.secret-key}")
-    private String secretKey;
-
-    @Value("${cloud.aws.s3.endpoint}")
-    private String endPoint;
-
-    @Value("${cloud.aws.region.static}")
-    private String regionName;
-
-    /**
-     * Provides an interface for accessing the Amazon S3 or Naver Cloud Object Storage
-     */
-    public AmazonS3 getAmazonS3() {
-        return AmazonS3ClientBuilder.standard()
-                .withEndpointConfiguration(new AwsClientBuilder.EndpointConfiguration(endPoint, regionName))
-                .withCredentials(new AWSStaticCredentialsProvider(new BasicAWSCredentials(accessKey, secretKey)))
-                .build();
-    }
-}
