+++ src/main/java/website/motus/domain/video/entity/Category.java
@@ -1,0 +1,5 @@
+package website.motus.domain.video.entity;
+
+public enum Category {
+    ARM, SHOULDER, KNEE, THIGH
+}
