+++ src/main/java/website/motus/domain/video/entity/Position.java
@@ -1,0 +1,5 @@
+package website.motus.domain.video.entity;
+
+public enum Position {
+    STANDING, SITTING, LYING
+}
