+++ src/main/java/website/motus/domain/video/repository/VideoRepository.java
@@ -1,0 +1,8 @@
+package website.motus.domain.video.repository;
+
+import website.motus.domain.video.entity.Video;
+import website.motus.domain.video.service.videoSearch.VideoSearch;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface VideoRepository extends JpaRepository<Video, Long>, VideoSearch {
+}
