+++ src/main/java/website/motus/global/baseEntity/BaseTimeEntity.java
@@ -1,0 +1,28 @@
+package website.motus.global.baseEntity;
+
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+import org.springframework.format.annotation.DateTimeFormat;
+
+import javax.persistence.Column;
+import javax.persistence.EntityListeners;
+import javax.persistence.MappedSuperclass;
+import java.time.LocalDateTime;
+
+@Getter
+@MappedSuperclass // When entities inherit, the class fields are also recognized as columns.
+@EntityListeners(value = { AuditingEntityListener.class }) // add Audting
+public class BaseTimeEntity {
+
+    @CreatedDate
+    @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
+    @Column(name = "regdate", updatable = false)
+    private LocalDateTime regDate;
+
+    @LastModifiedDate
+    @DateTimeFormat(pattern = "yyyy-MM-dd HH:mm:ss")
+    @Column(name ="moddate" )
+    private LocalDateTime modDate;
+}
