+++ src/main/java/website/motus/global/config/WebConfig.java
@@ -1,0 +1,18 @@
+package website.motus.global.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+                .allowedOrigins("*")
+                .allowedMethods("HEAD", "GET", "POST", "PUT", "DELETE", "PATCH")
+                .allowedHeaders("*");
+    }
+}
