--- src/test/java/com/hallym/rehab/RehabApplicationTests.java
@@ -1,15 +1,0 @@
-package com.hallym.rehab;
-
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@Disabled
-//@SpringBootTest
-class RehabApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
