+++ src/test/java/website/motus/RehabApplicationTests.java
@@ -1,0 +1,14 @@
+package website.motus;
+
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+
+@Disabled
+//@SpringBootTest
+class MotusApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
